requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
jinja2==3.1.6
httpx==0.28.1
python-dotenv==1.2.2
numpy==2.4.4
faiss-cpu==1.13.2
openai==2.32.0